refactor: update dir/file names (#35)
