+++ examples/todo-app/todo-example.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react'
 import { create } from 'zustand'
 import { zusound } from '../../packages' // Adjust the import path if necessary
-import { CodeViewer } from '../CodeViewer' // Import the CodeViewer
-import todoSource from './Todo.tsx?raw' // Import raw source code
+import { CodeViewer } from '../code-viewer' // Import the CodeViewer
+import todoSource from './todo-example.tsx?raw' // Import raw source code
 
 // --- Types ---
 interface Todo {
@@ -217,9 +217,8 @@ function TodoItem({ todo }: TodoItemProps) {
 
   return (
     <li
-      className={`flex items-center py-3 border-b border-gray-200 last:border-b-0 ${
-        todo.completed ? 'opacity-60' : ''
-      }`}
+      className={`flex items-center py-3 border-b border-gray-200 last:border-b-0 ${todo.completed ? 'opacity-60' : ''
+        }`}
     >
       <input
         type="checkbox"
@@ -299,15 +298,13 @@ function Controls() {
   }
 
   const getFilterButtonStyle = (controlFilter: Filter) =>
-    `btn text-xs sm:text-sm ${
-      filter === controlFilter ? 'btn-primary' : 'btn-outline hover:bg-gray-100'
+    `btn text-xs sm:text-sm ${filter === controlFilter ? 'btn-primary' : 'btn-outline hover:bg-gray-100'
     }`
 
   const getSortButtonStyle = (controlOrder: SortOrder) =>
-    `btn text-xs sm:text-sm ${
-      sortOrder === controlOrder && !isSorting
-        ? 'btn-secondary' // Indicate active sort order when not currently sorting
-        : 'btn-outline hover:bg-gray-100'
+    `btn text-xs sm:text-sm ${sortOrder === controlOrder && !isSorting
+      ? 'btn-secondary' // Indicate active sort order when not currently sorting
+      : 'btn-outline hover:bg-gray-100'
     } ${isSorting && sortOrder === controlOrder ? 'bg-yellow-200 text-yellow-800 border-yellow-300 animate-pulse' : ''}`
 
   return (
